Bug 1460489 - Part 4: Move IfEmitter to IfEmitter.{cpp.h}. r=jwalden
