+++ js/src/frontend/BytecodeEmitter.cpp
@@ -31,6 +31,7 @@
 #include "ds/Nestable.h"
 #include "frontend/BytecodeControlStructures.h"
 #include "frontend/EmitterScope.h"
+#include "frontend/IfEmitter.h"
 #include "frontend/Parser.h"
 #include "frontend/TDZCheckCache.h"
 #include "frontend/TokenStream.h"
@@ -522,367 +523,6 @@ class MOZ_STACK_CLASS TryEmitter
 #endif
         return true;
     }
-};
-
-// Class for emitting bytecode for blocks like if-then-else.
-//
-// This class can be used to emit single if-then-else block, or cascading
-// else-if blocks.
-//
-// Usage: (check for the return value is omitted for simplicity)
-//
-//   `if (cond) then_block`
-//     IfEmitter ifThen(this);
-//     emit(cond);
-//     ifThen.emitThen();
-//     emit(then_block);
-//     ifThen.emitEnd();
-//
-//   `if (cond) then_block else else_block`
-//     IfEmitter ifThenElse(this);
-//     emit(cond);
-//     ifThenElse.emitThenElse();
-//     emit(then_block);
-//     ifThenElse.emitElse();
-//     emit(else_block);
-//     ifThenElse.emitEnd();
-//
-//   `if (c1) b1 else if (c2) b2 else if (c3) b3 else b4`
-//     IfEmitter ifThenElse(this);
-//     emit(c1);
-//     ifThenElse.emitThenElse();
-//     emit(b1);
-//     ifThenElse.emitElseIf();
-//     emit(c2);
-//     ifThenElse.emitThenElse();
-//     emit(b2);
-//     ifThenElse.emitElseIf();
-//     emit(c3);
-//     ifThenElse.emitThenElse();
-//     emit(b3);
-//     ifThenElse.emitElse();
-//     emit(b4);
-//     ifThenElse.emitEnd();
-//
-//   `cond ? then_expr : else_expr`
-//     IfEmitter condElse(this);
-//     emit(cond);
-//     condElse.emitCond();
-//     emit(then_block);
-//     condElse.emitElse();
-//     emit(else_block);
-//     condElse.emitEnd();
-//
-class MOZ_STACK_CLASS IfEmitter
-{
-  public:
-    // Whether the then-clause, the else-clause, or else-if condition may
-    // contain declaration or access to lexical variables, which means they
-    // should have their own TDZCheckCache.  Basically TDZCheckCache should be
-    // created for each basic block, which then-clause, else-clause, and
-    // else-if condition are, but for internally used branches which are
-    // known not to touch lexical variables we can skip creating TDZCheckCache
-    // for them.
-    //
-    // See the comment for TDZCheckCache class for more details.
-    enum class Kind {
-        // For syntactic branches (if, if-else, and conditional expression),
-        // which basically may contain declaration or accesses to lexical
-        // variables inside then-clause, else-clause, and else-if condition.
-        MayContainLexicalAccessInBranch,
-
-        // For internally used branches which don't touch lexical variables
-        // inside then-clause, else-clause, nor else-if condition.
-        NoLexicalAccessInBranch
-    };
-
-  private:
-    BytecodeEmitter* bce_;
-
-    // Jump around the then clause, to the beginning of the else clause.
-    JumpList jumpAroundThen_;
-
-    // Jump around the else clause, to the end of the entire branch.
-    JumpList jumpsAroundElse_;
-
-    // The stack depth before emitting the then block.
-    // Used for restoring stack depth before emitting the else block.
-    // Also used for assertion to make sure then and else blocks pushed the
-    // same number of values.
-    int32_t thenDepth_;
-
-    Kind kind_;
-    Maybe<TDZCheckCache> tdzCache_;
-
-#ifdef DEBUG
-    // The number of values pushed in the then and else blocks.
-    int32_t pushed_;
-    bool calculatedPushed_;
-
-    // The state of this emitter.
-    //
-    // +-------+   emitCond +------+ emitElse +------+        emitEnd +-----+
-    // | Start |-+--------->| Cond |--------->| Else |------>+------->| End |
-    // +-------+ |          +------+          +------+       ^        +-----+
-    //           |                                           |
-    //           v emitThen +------+                         |
-    //        +->+--------->| Then |------------------------>+
-    //        ^  |          +------+                         ^
-    //        |  |                                           |
-    //        |  |                                           +---+
-    //        |  |                                               |
-    //        |  | emitThenElse +----------+   emitElse +------+ |
-    //        |  +------------->| ThenElse |-+--------->| Else |-+
-    //        |                 +----------+ |          +------+
-    //        |                              |
-    //        |                              | emitElseIf +--------+
-    //        |                              +----------->| ElseIf |-+
-    //        |                                           +--------+ |
-    //        |                                                      |
-    //        +------------------------------------------------------+
-    enum class State {
-        // The initial state.
-        Start,
-
-        // After calling emitThen.
-        Then,
-
-        // After calling emitCond.
-        Cond,
-
-        // After calling emitThenElse.
-        ThenElse,
-
-        // After calling emitElse.
-        Else,
-
-        // After calling emitElseIf.
-        ElseIf,
-
-        // After calling emitEnd.
-        End
-    };
-    State state_;
-#endif
-
-  protected:
-    // For InternalIfEmitter.
-    IfEmitter(BytecodeEmitter* bce, Kind kind)
-      : bce_(bce)
-      , thenDepth_(0)
-      , kind_(kind)
-#ifdef DEBUG
-      , pushed_(0)
-      , calculatedPushed_(false)
-      , state_(State::Start)
-#endif
-    {}
-
-  public:
-    explicit IfEmitter(BytecodeEmitter* bce)
-      : IfEmitter(bce, Kind::MayContainLexicalAccessInBranch)
-    {}
-
-  private:
-    MOZ_MUST_USE bool emitIfInternal(SrcNoteType type) {
-        MOZ_ASSERT_IF(state_ == State::ElseIf, tdzCache_.isSome());
-        MOZ_ASSERT_IF(state_ != State::ElseIf, tdzCache_.isNothing());
-
-        // The end of TDZCheckCache for cond for else-if.
-        if (kind_ == Kind::MayContainLexicalAccessInBranch)
-            tdzCache_.reset();
-
-        // Emit an annotated branch-if-false around the then part.
-        if (!bce_->newSrcNote(type))
-            return false;
-        if (!bce_->emitJump(JSOP_IFEQ, &jumpAroundThen_))
-            return false;
-
-        // To restore stack depth in else part, save depth of the then part.
-#ifdef DEBUG
-        // If DEBUG, this is also necessary to calculate |pushed_|.
-        thenDepth_ = bce_->stackDepth;
-#else
-        if (type == SRC_COND || type == SRC_IF_ELSE)
-            thenDepth_ = bce_->stackDepth;
-#endif
-
-        // Enclose then-branch with TDZCheckCache.
-        if (kind_ == Kind::MayContainLexicalAccessInBranch)
-            tdzCache_.emplace(bce_);
-
-        return true;
-    }
-
-    void calculateOrCheckPushed() {
-#ifdef DEBUG
-        if (!calculatedPushed_) {
-            pushed_ = bce_->stackDepth - thenDepth_;
-            calculatedPushed_ = true;
-        } else {
-            MOZ_ASSERT(pushed_ == bce_->stackDepth - thenDepth_);
-        }
-#endif
-    }
-
-  public:
-    MOZ_MUST_USE bool emitThen() {
-        MOZ_ASSERT(state_ == State::Start || state_ == State::ElseIf);
-        if (!emitIfInternal(SRC_IF))
-            return false;
-
-#ifdef DEBUG
-        state_ = State::Then;
-#endif
-        return true;
-    }
-
-    MOZ_MUST_USE bool emitCond() {
-        MOZ_ASSERT(state_ == State::Start);
-        if (!emitIfInternal(SRC_COND))
-            return false;
-
-#ifdef DEBUG
-        state_ = State::Cond;
-#endif
-        return true;
-    }
-
-    MOZ_MUST_USE bool emitThenElse() {
-        MOZ_ASSERT(state_ == State::Start || state_ == State::ElseIf);
-        if (!emitIfInternal(SRC_IF_ELSE))
-            return false;
-
-#ifdef DEBUG
-        state_ = State::ThenElse;
-#endif
-        return true;
-    }
-
-  private:
-    MOZ_MUST_USE bool emitElseInternal() {
-        calculateOrCheckPushed();
-
-        // The end of TDZCheckCache for then-clause.
-        if (kind_ == Kind::MayContainLexicalAccessInBranch) {
-            MOZ_ASSERT(tdzCache_.isSome());
-            tdzCache_.reset();
-        }
-
-        // Emit a jump from the end of our then part around the else part. The
-        // patchJumpsToTarget call at the bottom of this function will fix up
-        // the offset with jumpsAroundElse value.
-        if (!bce_->emitJump(JSOP_GOTO, &jumpsAroundElse_))
-            return false;
-
-        // Ensure the branch-if-false comes here, then emit the else.
-        if (!bce_->emitJumpTargetAndPatch(jumpAroundThen_))
-            return false;
-
-        // Clear jumpAroundThen_ offset, to tell emitEnd there was an else part.
-        jumpAroundThen_ = JumpList();
-
-        // Restore stack depth of the then part.
-        bce_->stackDepth = thenDepth_;
-#ifdef DEBUG
-        state_ = State::Else;
-#endif
-        return true;
-    }
-
-  public:
-    MOZ_MUST_USE bool emitElse() {
-        MOZ_ASSERT(state_ == State::ThenElse || state_ == State::Cond);
-
-        if (!emitElseInternal())
-            return false;
-
-        // Enclose else-branch with TDZCheckCache.
-        if (kind_ == Kind::MayContainLexicalAccessInBranch)
-            tdzCache_.emplace(bce_);
-
-#ifdef DEBUG
-        state_ = State::Else;
-#endif
-        return true;
-    }
-
-    MOZ_MUST_USE bool emitElseIf() {
-        MOZ_ASSERT(state_ == State::ThenElse);
-
-        if (!emitElseInternal())
-            return false;
-
-        // Enclose cond for else-if with TDZCheckCache.
-        if (kind_ == Kind::MayContainLexicalAccessInBranch)
-            tdzCache_.emplace(bce_);
-
-#ifdef DEBUG
-        state_ = State::ElseIf;
-#endif
-        return true;
-    }
-
-    MOZ_MUST_USE bool emitEnd() {
-        MOZ_ASSERT(state_ == State::Then || state_ == State::Else);
-        // If there was an else part for the last branch, jumpAroundThen_ is
-        // already fixed up when emitting the else part.
-        MOZ_ASSERT_IF(state_ == State::Then, jumpAroundThen_.offset != -1);
-        MOZ_ASSERT_IF(state_ == State::Else, jumpAroundThen_.offset == -1);
-
-        // The end of TDZCheckCache for then or else-clause.
-        if (kind_ == Kind::MayContainLexicalAccessInBranch) {
-            MOZ_ASSERT(tdzCache_.isSome());
-            tdzCache_.reset();
-        }
-
-        calculateOrCheckPushed();
-
-        if (jumpAroundThen_.offset != -1) {
-            // No else part for the last branch, fixup the branch-if-false to
-            // come here.
-            if (!bce_->emitJumpTargetAndPatch(jumpAroundThen_))
-                return false;
-        }
-
-        // Patch all the jumps around else parts.
-        if (!bce_->emitJumpTargetAndPatch(jumpsAroundElse_))
-            return false;
-
-#ifdef DEBUG
-        state_ = State::End;
-#endif
-        return true;
-    }
-
-#ifdef DEBUG
-    // Returns the number of values pushed onto the value stack inside
-    // `then_block` and `else_block`.
-    // Can be used in assertion after emitting if-then-else.
-    int32_t pushed() const {
-        return pushed_;
-    }
-
-    // Returns the number of values popped onto the value stack inside
-    // `then_block` and `else_block`.
-    // Can be used in assertion after emitting if-then-else.
-    int32_t popped() const {
-        return -pushed_;
-    }
-#endif
-};
-
-// Class for emitting bytecode for blocks like if-then-else which doesn't touch
-// lexical variables.
-//
-// See the comments above NoLexicalAccessInBranch for more details when to use
-// this instead of IfEmitter.
-class MOZ_STACK_CLASS InternalIfEmitter : public IfEmitter
-{
-  public:
-    explicit InternalIfEmitter(BytecodeEmitter* bce)
-      : IfEmitter(bce, Kind::NoLexicalAccessInBranch)
-    {}
 };
 
 // OOOEEE start
